+++ ethclient/customclient/ethclient_ext.go
@@ -1,9 +1,10 @@
 // (c) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package corethclient
+package customclient
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 
@@ -12,9 +13,36 @@ import (
 
 	"github.com/ava-labs/coreth/ethclient"
 	"github.com/ava-labs/coreth/plugin/evm/customtypes"
+	"github.com/ava-labs/coreth/rpc"
 )
 
 var _ ethclient.BlockHook = (*extBlockHook)(nil)
+
+// Client wraps the ethclient.Client interface to provide extra data types (in header, block body).
+// If you want to use the standardized Ethereum RPC functionality without extra types, use ethclient.Client instead.
+type Client struct {
+	c *rpc.Client
+	ethclient.Client
+}
+
+// New creates a client that uses the given RPC client.
+func New(c *rpc.Client) *Client {
+	return &Client{c: c, Client: ethclient.NewClientWithHook(c, &extBlockHook{})}
+}
+
+// Dial connects a client to the given URL.
+func Dial(rawurl string) (*Client, error) {
+	return DialContext(context.Background(), rawurl)
+}
+
+// DialContext connects a client to the given URL with context.
+func DialContext(ctx context.Context, rawurl string) (*Client, error) {
+	c, err := rpc.DialContext(ctx, rawurl)
+	if err != nil {
+		return nil, err
+	}
+	return New(c), nil
+}
 
 // extBlockHook is a hook that is called when a block is decoded.
 type extBlockHook struct{}
